refactor filter into it's own component
