+++ app/client/components/helpers/responsive.tsx
@@ -1,5 +1,5 @@
 import * as React from "react";
-import MediaQuery from "react-responsive";
+import MediaQuery, { useMediaQuery } from "react-responsive";
 
 type Props = {
   children: React.ReactNode;
@@ -12,6 +12,11 @@ export const Mobile: React.FunctionComponent<Props> = ({
 }) => {
   return <MediaQuery maxWidth={767}>{children}</MediaQuery>;
 };
+
+export const isMobile = () => {
+  return useMediaQuery({ query: "(max-width: 767px)" });
+};
+
 export const Default: React.FunctionComponent<Props> = ({
   children
 }: {
